ch5/ep2/exercises/requirements.txt
numpy==1.20.2
pandas==1.2.4
SQLAlchemy==1.4.13
tabulate==0.8.9
jupyterlab==3.0.14
notebook==6.3.0
jupyterthemes==0.20.0
pyyaml==5.4.1
colored==1.4.3